Added monster to the Main Scene.
For some reason, the raycast is unable to see the player.  It is working, but it cannot transition into the chase state due to this.

Also seeded the random function for the AI's waypoint selection based on the device time in milliseconds, and made slight adjustments to the ongoing saving and loading project.