Added the fisherman npc.
Unsure why, but in scenes with multiple npc characters, some of them will skip the first line of their dialogue.  I have no idea why this is happening (I've been looking for an explanation).  Once the scene is loaded, setting any character to be inactive and then active again is the only way I've found to overcome this.