All the npcs now have dialgue.
They have been set up in the park and dock scenes.
I still don't know why the bug is happening, but have found a workaround that will maybe work for us.
Am planning to fix the jogger states (it keeps jogging when you talk to it until it reaches the next waypoint)